+++ src/Module/Widget/PartPayment/Html.php
@@ -29,16 +29,12 @@ use Resursbank\Ecom\Lib\Locale\Translator;
 use Resursbank\Ecom\Lib\Model\AnnuityFactor\AnnuityInformation;
 use Resursbank\Ecom\Lib\Model\PaymentMethod;
 use Resursbank\Ecom\Lib\Model\PriceSignage\Cost;
-use Resursbank\Ecom\Lib\UserSettings\Field;
 use Resursbank\Ecom\Lib\Utilities\Price;
 use Resursbank\Ecom\Lib\Widget\Widget;
 use Resursbank\Ecom\Module\AnnuityFactor\Repository;
-use Resursbank\Ecom\Module\PriceSignage\Repository as SignageRepository;
-use Resursbank\Ecom\Module\UserSettings\Repository as UserSettingsRepository;
 use Resursbank\Ecom\Module\Widget\ConsumerCreditWarning\Html as ConsumerCreditWarning;
 use Resursbank\Ecom\Module\Widget\PartPayment\Traits\Common;
 use Resursbank\Ecom\Module\Widget\ReadMore\Html as ReadMoreHtml;
-use Resursbank\Woocommerce\Util\Log;
 use Throwable;
 
 use function max;
@@ -243,7 +239,7 @@ class Html extends Widget
 
         try {
             $result = sprintf(
-                Translator::translate('rb-pp-not-eligible'),
+                Translator::translate(phraseId: 'rb-pp-not-eligible'),
                 $period
             );
         } catch (Throwable $e) {