+++ tests/Integration/Module/Payment/Api/RefundTest.php
@@ -1,11 +1,11 @@
 <?php
 
-/** @noinspection EfferentObjectCouplingInspection */
-
 /**
  * Copyright © Resurs Bank AB. All rights reserved.
  * See LICENSE for license details.
  */
+
+/** @noinspection EfferentObjectCouplingInspection */
 
 declare(strict_types=1);